Unshadow RSA public key
Due to future binary compatibility the RSA public key structure is visible all the time no matter if ROW_RSA_ENABLED is defined or not. When macro ROW_RSA_ENABLED is not defined the RSA signature verification is still shadowed and addkey action will fail when trying to add WA RSA public key.